[refactory] PatientService e AttendanceService
