--- src/main/java/br/com/nivlabs/gp/service/AttendanceService.java
@@ -1,370 +1,0 @@
-package br.com.nivlabs.gp.service;
-
-import java.time.LocalDateTime;
-import java.util.ArrayList;
-import java.util.List;
-import java.util.stream.Collectors;
-
-import org.slf4j.Logger;
-import org.springframework.beans.BeanUtils;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
-import org.springframework.http.HttpStatus;
-import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.stereotype.Service;
-
-import br.com.nivlabs.gp.config.security.UserOfSystem;
-import br.com.nivlabs.gp.controller.filters.AttendanceFilters;
-import br.com.nivlabs.gp.enums.DocumentType;
-import br.com.nivlabs.gp.enums.EntryType;
-import br.com.nivlabs.gp.enums.EventType;
-import br.com.nivlabs.gp.exception.HttpException;
-import br.com.nivlabs.gp.models.domain.Accommodation;
-import br.com.nivlabs.gp.models.domain.Attendance;
-import br.com.nivlabs.gp.models.domain.AttendanceEvent;
-import br.com.nivlabs.gp.models.domain.Patient;
-import br.com.nivlabs.gp.models.domain.PatientAllergy;
-import br.com.nivlabs.gp.models.domain.Person;
-import br.com.nivlabs.gp.models.domain.Responsible;
-import br.com.nivlabs.gp.models.dto.AccommodationDTO;
-import br.com.nivlabs.gp.models.dto.AttendanceDTO;
-import br.com.nivlabs.gp.models.dto.AttendanceEventDTO;
-import br.com.nivlabs.gp.models.dto.CloseAttandenceDTO;
-import br.com.nivlabs.gp.models.dto.DocumentDTO;
-import br.com.nivlabs.gp.models.dto.EvolutionInfoDTO;
-import br.com.nivlabs.gp.models.dto.MedicalRecordDTO;
-import br.com.nivlabs.gp.models.dto.MedicineInfoDTO;
-import br.com.nivlabs.gp.models.dto.NewAttandenceDTO;
-import br.com.nivlabs.gp.models.dto.NewAttendanceEventDTO;
-import br.com.nivlabs.gp.models.dto.PatientInfoDTO;
-import br.com.nivlabs.gp.models.dto.ResponsibleInfoDTO;
-import br.com.nivlabs.gp.models.dto.UserInfoDTO;
-import br.com.nivlabs.gp.repository.AttendanceEventRepository;
-import br.com.nivlabs.gp.repository.AttendanceRepository;
-import br.com.nivlabs.gp.service.patient.PatientService;
-
-/**
- * Classe AttendanceService.java
- *
- * @author <a href="carolexc@gmail.com">Caroline Aguiar</a>
- * @since 8 de set de 2019
- */
-@Service
-public class AttendanceService implements BaseService {
-
-    @Autowired
-    private Logger logger;
-    @Autowired
-    private AttendanceRepository dao;
-    @Autowired
-    private AttendanceEventRepository attendanceEventRepo;
-    @Autowired
-    private PatientService patientService;
-    @Autowired
-    private ResponsibleService responsibleService;
-    @Autowired
-    private UserService userService;
-    @Autowired
-    private AttendanceEventService attendanceEventService;
-
-    public Page<AttendanceDTO> getAttendancesPage(AttendanceFilters filters, Pageable pageRequest) {
-        return dao.resumedList(filters, pageRequest);
-    }
-
-    /**
-     * Busca histórico de Atendimentos por Paciente
-     *
-     * @param patientId
-     * @return lista de AttendanceDTO
-     */
-    public List<AttendanceDTO> getAttandenceByPatientId(Long patientId) {
-        List<Attendance> list = dao.findByPatient(new Patient(patientId));
-
-        if (list.isEmpty())
-            throw new HttpException(HttpStatus.NOT_FOUND,
-                    String.format("Não existe atendimento para o paciente %s", patientId));
-
-        List<AttendanceDTO> listOfDto = new ArrayList<>();
-        list.forEach(attendance -> {
-            AttendanceDTO dto = new AttendanceDTO();
-            dto.setId(attendance.getId());
-            dto.setEntryDatetime(attendance.getEntryDateTime());
-            dto.setEntryCause(attendance.getReasonForEntry());
-            dto.setIsFinished(Boolean.valueOf(attendance.getExitDateTime() != null));
-            listOfDto.add(dto);
-        });
-        return listOfDto;
-    }
-
-    /**
-     * Busca Prontuário do atendimento por Id
-     *
-     * @param id
-     * @return MedicalRecordDTO
-     */
-    public MedicalRecordDTO findMedicalRecordByAttendanceId(Long id) {
-        logger.info("Verificando se o atendimento informado existe. Atendimento :: {}", id);
-        Attendance objectFromDb = dao.findById(id).orElseThrow(() -> new HttpException(HttpStatus.NOT_FOUND,
-                String.format("Prontuário com código %s não encontrado", id)));
-        Person person = objectFromDb.getPatient().getPerson();
-        MedicalRecordDTO medicalRecord = new MedicalRecordDTO();
-        BeanUtils.copyProperties(person, medicalRecord);
-        medicalRecord.setId(objectFromDb.getId());
-        medicalRecord.setPatientId(objectFromDb.getPatient().getId());
-        medicalRecord.setDocument(new DocumentDTO(null, DocumentType.CPF, person.getCpf(), null, null, null, null));
-        medicalRecord.setCnsNumber(objectFromDb.getPatient().getCnsNumber());
-        medicalRecord.setEntryDateTime(objectFromDb.getEntryDateTime());
-        medicalRecord.setExitDateTime(objectFromDb.getExitDateTime());
-        medicalRecord.setAttendanceLevel(objectFromDb.getLevel());
-        medicalRecord.setReasonForEntry(objectFromDb.getReasonForEntry());
-
-        processEvents(objectFromDb, medicalRecord);
-
-        if (!medicalRecord.getEvents().isEmpty())
-            medicalRecord.setLastAccommodation(getLastAccommodationByPatientId(medicalRecord.getEvents()));
-        medicalRecord.getAllergies()
-                .addAll(objectFromDb.getPatient().getAllergies().stream().map(PatientAllergy::getDescription).collect(Collectors.toList()));
-        return medicalRecord;
-    }
-
-    /**
-     * Processa eventos do atendimento no prontuário
-     * 
-     * @param objectFromDb
-     * @param medicalRecord
-     */
-    private void processEvents(Attendance objectFromDb, MedicalRecordDTO medicalRecord) {
-        objectFromDb.getEvents().forEach(event -> {
-            processEvent(medicalRecord, event);
-        });
-    }
-
-    /**
-     * Processa evento no prontuário
-     * 
-     * @param medicalRecord
-     * @param event
-     */
-    private void processEvent(MedicalRecordDTO medicalRecord, AttendanceEvent event) {
-        medicalRecord.getEvents().add(event.getDTO());
-        if (event.getEventType() == EventType.EVOLUTION) {
-            processEvolution(medicalRecord, event);
-        } else if (event.getEventType() == EventType.MEDICINE) {
-            processMedications(medicalRecord, event);
-        }
-    }
-
-    /**
-     * Processa medicamentos no protuário
-     * 
-     * @param medicalRecord
-     * @param event
-     */
-    private void processMedications(MedicalRecordDTO medicalRecord, AttendanceEvent event) {
-        MedicineInfoDTO medicine = new MedicineInfoDTO();
-        medicine.setId(event.getId());
-        medicine.setDatetime(event.getEventDateTime());
-        medicine.setDescription(event.getTitle());
-        medicine.setAmount(event.getObservations());
-        medicine.setResponsibleForTheAdministration(event.getResponsible().getPerson().getFullName());
-
-        medicalRecord.getMedicines().add(medicine);
-    }
-
-    /**
-     * Processa evoluções no prontuário
-     * 
-     * @param medicalRecord
-     * @param event
-     */
-    private void processEvolution(MedicalRecordDTO medicalRecord, AttendanceEvent event) {
-        EvolutionInfoDTO evolution = new EvolutionInfoDTO();
-        evolution.setId(event.getId());
-        evolution.setAccommodationId(medicalRecord.getLastAccommodation() != null ? medicalRecord.getLastAccommodation().getId() : null);
-        evolution.setAttendanceId(medicalRecord.getId());
-        evolution.setDescription("EVOLUÇÃO");
-        evolution.setDatetime(event.getEventDateTime());
-        evolution.setResponsibleName(event.getResponsible().getPerson().getFullName());
-
-        medicalRecord.getEvolutions().add(evolution);
-    }
-
-    public Attendance persist(Attendance entity) {
-        entity.setId(null);
-        return dao.save(entity);
-    }
-
-    /**
-     * Realiza a criação de uma visita de paciente à partir do DTO
-     *
-     * @param NewAttandenceDTO Novo atendimento
-     * @return MedicalRecordDTO
-     */
-    public MedicalRecordDTO persistNewAttendance(NewAttandenceDTO request) {
-        MedicalRecordDTO attendance = null;
-        try {
-            getActiveMedicalRecord(request.getPatientId());
-        } catch (HttpException e) {
-            if (e.getStatus().equals(HttpStatus.UNPROCESSABLE_ENTITY)) {
-                PatientInfoDTO savedPatient = patientService.findByPatientId(request.getPatientId());
-
-                Attendance convertedAttendance = new Attendance();
-                convertedAttendance.setReasonForEntry(request.getEntryCause());
-                convertedAttendance.setLevel(request.getLevel());
-                convertedAttendance.setPatient(new Patient(savedPatient.getId()));
-                convertedAttendance.setCurrentAccommodation(new Accommodation(request.getAccommodationId(), null, null, null));
-
-                switch (request.getEventType()) {
-                    case MEDICAL_EMERGENCY:
-                        convertedAttendance.setEntryType(EntryType.EMERGENCY);
-                        break;
-                    case MEDICAL_APPOINTMENT:
-                        convertedAttendance.setEntryType(EntryType.CLINICAL);
-                        break;
-                    case REMOTE_MEDICAL_APPOINTMENT:
-                        convertedAttendance.setEntryType(EntryType.REMOTE);
-                        break;
-                    default:
-                        convertedAttendance.setEntryType(EntryType.GENERIC);
-                }
-                convertedAttendance = persist(convertedAttendance);
-                createEntryEvent(convertedAttendance, request);
-                attendance = getActiveMedicalRecord(request.getPatientId());
-
-                return attendance;
-            }
-        }
-
-        throw new HttpException(HttpStatus.UNPROCESSABLE_ENTITY, String.format(
-                                                                               "O paciente de código %s já possui um atendimento ativo, favor realizar a alta do mesmo para iniciar um novo.",
-                                                                               request.getPatientId()));
-
-    }
-
-    /**
-     * Cria evento de entrada
-     *
-     * @param convertedAtendance
-     * @param request
-     */
-    private void createEntryEvent(Attendance convertedAtendance, NewAttandenceDTO request) {
-        AttendanceEvent entryEvent = new AttendanceEvent();
-        entryEvent.setEventDateTime(convertedAtendance.getEntryDateTime());
-        entryEvent.setTitle(convertedAtendance.getReasonForEntry());
-        entryEvent.setAttendance(new Attendance(convertedAtendance.getId()));
-        entryEvent.setAccommodation(convertedAtendance.getCurrentAccommodation());
-
-        if (request.getResponsibleId() != null)
-            // Verificar Responsável
-            entryEvent.setResponsible(new Responsible(request.getResponsibleId()));
-
-        entryEvent.setEventType(request.getEventType());
-
-        attendanceEventRepo.save(entryEvent);
-    }
-
-    /**
-     * Realiza a busca de um prontuário ativo por código de paciente
-     *
-     * @return MedicalRecordDTO
-     */
-    public MedicalRecordDTO getActiveMedicalRecord(Long patientId) {
-        PatientInfoDTO patient = patientService.findByPatientId(patientId);
-        Attendance attendanceFromDb = dao.findByPatientAndExitDateTimeIsNull(new Patient(patient.getId()))
-                .orElseThrow(() -> new HttpException(HttpStatus.UNPROCESSABLE_ENTITY, String.format(
-                                                                                                    "Nenhum atendimento ativo encontrado para %s, inicie um novo atendimento para o paciente.",
-                                                                                                    patient.getFullName())));
-        MedicalRecordDTO medicalRecord = new MedicalRecordDTO();
-        BeanUtils.copyProperties(patient, medicalRecord);
-        medicalRecord.setId(attendanceFromDb.getId());
-        medicalRecord.setPatientId(attendanceFromDb.getPatient().getId());
-        medicalRecord
-                .setDocument(new DocumentDTO(null, patient.getDocument().getType(), patient.getDocument().getValue(), null, null, null,
-                        null));
-        medicalRecord.setEntryDateTime(attendanceFromDb.getEntryDateTime());
-        medicalRecord.setExitDateTime(attendanceFromDb.getExitDateTime());
-        medicalRecord.setAttendanceLevel(attendanceFromDb.getLevel());
-        medicalRecord.setReasonForEntry(attendanceFromDb.getReasonForEntry());
-
-        processEvents(attendanceFromDb, medicalRecord);
-
-        if (!medicalRecord.getEvents().isEmpty())
-            medicalRecord.setLastAccommodation(getLastAccommodationByPatientId(medicalRecord.getEvents()));
-        return medicalRecord;
-    }
-
-    /**
-     * Pega a última acomodação do paciente baseada nos eventos do prontuário
-     *
-     * @param listOfEventsFromDb
-     * @return
-     */
-    private AccommodationDTO getLastAccommodationByPatientId(List<AttendanceEventDTO> listOfEventsFromDb) {
-        return listOfEventsFromDb.get(listOfEventsFromDb.size() - 1).getAccommodation();
-    }
-
-    /**
-     * Realiza alta de paciente com atendimento ativo
-     *
-     * @param id
-     * @param request
-     */
-    public void closeAttendance(Long id, CloseAttandenceDTO request) {
-        logger.info("Iniciando processo de encerramento de atendimento,\nVerificando situação do atendimento solicitado...");
-        Attendance attendanceFromDb = dao.findById(id).orElseThrow(() -> new HttpException(HttpStatus.UNPROCESSABLE_ENTITY,
-                "Não é possível encerrar um atendimento inexistente"));
-        if (attendanceFromDb.getExitDateTime() != null) {
-            throw new HttpException(HttpStatus.UNPROCESSABLE_ENTITY, "Este atendimento já foi encerrado");
-        }
-        NewAttendanceEventDTO newAttendanceEvent = createAttendanceFromCloseRequest(attendanceFromDb, request);
-        attendanceFromDb.setExitDateTime(newAttendanceEvent.getEventDateTime());
-        dao.save(attendanceFromDb);
-    }
-
-    /**
-     * Cria um evento de encerramento de atendimento (Alta)
-     *
-     * @param attendance
-     * @param request
-     * @return
-     */
-    private NewAttendanceEventDTO createAttendanceFromCloseRequest(Attendance attendance, CloseAttandenceDTO request) {
-        logger.info("Criando evento de atendimento para alta de paciente");
-        NewAttendanceEventDTO event = new NewAttendanceEventDTO();
-        event.setAttendanceId(attendance.getId());
-        event.setEventDateTime(LocalDateTime.now());
-        event.setEventType(request.getEventType());
-        event.setAccommodation(new AccommodationDTO(attendance.getCurrentAccommodation().getId()));
-        event.setResponsible(getResponsibleFromUserSession());
-        event.setObservations(request.getObservations());
-
-        attendanceEventService.persistNewAttendanceEvent(event, null);
-        logger.info("Evento de Atendimento para alta de paciente criado com sucesso!");
-
-        return event;
-    }
-
-    /**
-     * Busca o usuário logado na sessão
-     *
-     * @return
-     */
-    private ResponsibleInfoDTO getResponsibleFromUserSession() {
-        logger.info("Buscando usuário da sessão...");
-        String userName = ((UserOfSystem) SecurityContextHolder.getContext().getAuthentication().getPrincipal()).getUsername();
-        UserInfoDTO logedUser = userService.findByUserName(userName);
-
-        logger.info("Iniciando busca de responsável pelo usuário da requisição...");
-        ResponsibleInfoDTO responsibleInformations = responsibleService.findByCpf(logedUser.getDocument().getValue());
-        if (responsibleInformations.getId() == null)
-            throw new HttpException(HttpStatus.FORBIDDEN, "Sem presmissão! Você não tem um profissional vinculado ao seu usuário.");
-        logger.info("Profissional encontrado :: {}", responsibleInformations.getFullName());
-
-        logger.info("Realizando processamento do profissional para a requisição de criação de evento de atendimento");
-        ResponsibleInfoDTO responsible = new ResponsibleInfoDTO();
-        BeanUtils.copyProperties(responsibleInformations, responsible);
-        return responsible;
-    }
-
-}
