+++ src/main/java/br/com/nivlabs/gp/service/patient/PatientService.java
@@ -1,45 +1,39 @@
 package br.com.nivlabs.gp.service.patient;
 
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
 import br.com.nivlabs.gp.controller.filters.PatientFilters;
-import br.com.nivlabs.gp.models.domain.PatientAllergy;
 import br.com.nivlabs.gp.models.dto.PatientAllergiesDTO;
 import br.com.nivlabs.gp.models.dto.PatientDTO;
 import br.com.nivlabs.gp.models.dto.PatientInfoDTO;
-import br.com.nivlabs.gp.repository.PatientAllergyRepository;
 import br.com.nivlabs.gp.service.BaseService;
 import br.com.nivlabs.gp.service.patient.business.CreatePatientBusinessHandler;
 import br.com.nivlabs.gp.service.patient.business.SearchPatientBusinessHandler;
+import br.com.nivlabs.gp.service.patient.business.UpdatePatientAllergiesBusinessHandler;
 import br.com.nivlabs.gp.service.patient.business.UpdatePatientBusinessHandler;
 
 /**
  * 
- * Classe PatientService.java
+ * Camada de serviços com todos os processos existentes de manipulação do paciente
  *
- * @author <a href="carolexc@gmail.com">Caroline Aguiar</a>
+ * @author viniciosarodrigues
+ * @since 19-09-2021
  *
- * @since 15 de set de 2019
  */
 @Service
 public class PatientService implements BaseService {
 
-    private Logger logger = LoggerFactory.getLogger(this.getClass());
-
-    @Autowired
-    private PatientAllergyRepository patientAllergyDao;
-    
     @Autowired
     SearchPatientBusinessHandler patientSearchBusinessHandler;
     @Autowired
     CreatePatientBusinessHandler createPatientBusinessHandler;
     @Autowired
     UpdatePatientBusinessHandler updatePatientBusinessHandler;
+    @Autowired
+    UpdatePatientAllergiesBusinessHandler updatePatientAllergiesBusinessHandler;
 
     /**
      * Busca uma página de informações resumidas de pacientes
@@ -86,11 +80,11 @@ public class PatientService implements BaseService {
     }
 
     /**
-     * Atualiza informações do paciente
+     * Atualiza informações de um paciente já existente na base de dados
      * 
-     * @param id
-     * @param entity
-     * @return
+     * @param id Identificador único do paciente
+     * @param request Objeto de transferência com informações do paciente
+     * @return Informações atualizadas do paciente
      */
     public PatientInfoDTO update(Long id, PatientInfoDTO entity) {
         entity.setId(id);
@@ -110,22 +104,11 @@ public class PatientService implements BaseService {
     /**
      * Atualiza as alergias do paciente
      * 
-     * @param id
-     * @param allergies
+     * @param patientId Identificador único do paciente
+     * @param request Requisição com informações das alergias do paciente
      */
-    public void updateAllergies(Long patientId, PatientAllergiesDTO allergies) {
-        logger.info("Iniciando atualização de alergidas para o paciente :: {}", patientId);
-        logger.info("Limpando alergias anteriores...");
-        patientAllergyDao.deleteByPatientId(patientId);
-
-        if (allergies != null && !allergies.getDescriptions().isEmpty()) {
-            logger.info("Total de Alergias à processar no paciente :: {}", allergies.getDescriptions().size());
-            for (String allergy : allergies.getDescriptions()) {
-                patientAllergyDao.saveAndFlush(new PatientAllergy(patientId, allergy));
-            }
-        }
-        logger.info("Atualização das alergias do paciente concluída com sucesso!");
-
+    public void updateAllergies(Long patientId, PatientAllergiesDTO request) {
+        updatePatientAllergiesBusinessHandler.update(patientId, request);
     }
 
 }
